Add: useFetchRestaurants custom hook
